Relocate stage logics for k8s plugin
